Change actions to resource-based ones
