--- src/actions/account-selector.js
@@ -1,5 +1,0 @@
-export const ACTIVATE_ACCOUNT = 'ACTIVATE_ACCOUNT';
-
-export const activateAccount = (index) => {
-  return { type: ACTIVATE_ACCOUNT, index: parseInt(index) }
-}
