--- src/actions/editor.js
@@ -1,10 +1,0 @@
-export const SET_TEXT = 'SET_TEXT';
-export const CLEAR_TEXT = 'CLEAR_TEXT';
-
-export const clearText = () => {
-  return { type: CLEAR_TEXT }
-}
-
-export const setText = (text) => {
-  return { type: SET_TEXT, text }
-}
