+++ src/actions/index.js
@@ -1,15 +1,13 @@
-import * as accountSelector from './account-selector';
-import * as app             from './app';
-import * as editor          from './editor';
-import * as listSelector    from './list-selector';
-import * as timeline        from './timeline';
-import * as tweetTab        from './tweet-tab';
+import * as accounts from './accounts';
+import * as lists    from './lists';
+import * as tabs     from './tabs';
+import * as texts    from './texts';
+import * as tweets   from './tweets';
 
 export default Object.assign(
-  accountSelector,
-  app,
-  editor,
-  listSelector,
-  timeline,
-  tweetTab,
+  accounts,
+  lists,
+  tabs,
+  texts,
+  tweets,
 )
