+++ src/actions/tabs.js
